+++ packages/accounts-meetup/meetup.js
@@ -1,20 +1,19 @@
 Accounts.oauth.registerService('meetup');
 
 if (Meteor.isClient) {
-  const loginWithMeetup = function(options, callback) {
+  const loginWithMeetup = (options, callback) => {
     // support a callback without options
     if (! callback && typeof options === "function") {
       callback = options;
       options = null;
     }
 
-    var credentialRequestCompleteCallback = Accounts.oauth.credentialRequestCompleteHandler(callback);
+    const credentialRequestCompleteCallback = Accounts.oauth.credentialRequestCompleteHandler(callback);
     Meetup.requestCredential(options, credentialRequestCompleteCallback);
   };
   Accounts.registerClientLoginFunction('meetup', loginWithMeetup);
-  Meteor.loginWithMeetup = function () {
-    return Accounts.applyLoginFunction('meetup', arguments);
-  };
+  Meteor.loginWithMeetup = 
+    (...args) => Accounts.applyLoginFunction('meetup', args);
 } else {
   Accounts.addAutopublishFields({
     // publish all fields including access token, which can legitimately