+++ packages/accounts-ui-unstyled/login_buttons_dialogs.js
@@ -1,22 +1,23 @@
+import { displayName, dropdown, validatePassword } from './login_buttons.js';
 // for convenience
-var loginButtonsSession = Accounts._loginButtonsSession;
+const loginButtonsSession = Accounts._loginButtonsSession;
 
 // since we don't want to pass around the callback that we get from our event
 // handlers, we just make it a variable for the whole file
-var doneCallback;
+let doneCallback;
 
-Accounts.onResetPasswordLink(function (token, done) {
+Accounts.onResetPasswordLink((token, done) => {
   loginButtonsSession.set("resetPasswordToken", token);
   doneCallback = done;
 });
 
-Accounts.onEnrollmentLink(function (token, done) {
+Accounts.onEnrollmentLink((token, done) => {
   loginButtonsSession.set("enrollAccountToken", token);
   doneCallback = done;
 });
 
-Accounts.onEmailVerificationLink(function (token, done) {
-  Accounts.verifyEmail(token, function (error) {
+Accounts.onEmailVerificationLink((token, done) => {
+  Accounts.verifyEmail(token, error => {
     if (! error) {
       loginButtonsSession.set('justVerifiedEmail', true);
     }
@@ -32,29 +33,27 @@ Accounts.onEmailVerificationLink(function (token, done) {
 //
 
 Template._resetPasswordDialog.events({
-  'click #login-buttons-reset-password-button': function () {
-    resetPassword();
-  },
-  'keypress #reset-password-new-password': function (event) {
+  'click #login-buttons-reset-password-button': () => resetPassword(),
+  'keypress #reset-password-new-password': event => {
     if (event.keyCode === 13)
       resetPassword();
   },
-  'click #login-buttons-cancel-reset-password': function () {
+  'click #login-buttons-cancel-reset-password': () => {
     loginButtonsSession.set('resetPasswordToken', null);
     if (doneCallback)
       doneCallback();
   }
 });
 
-var resetPassword = function () {
+const resetPassword = () => {
   loginButtonsSession.resetMessages();
-  var newPassword = document.getElementById('reset-password-new-password').value;
+  const newPassword = document.getElementById('reset-password-new-password').value;
   if (!validatePassword(newPassword))
     return;
 
   Accounts.resetPassword(
     loginButtonsSession.get('resetPasswordToken'), newPassword,
-    function (error) {
+    error => {
       if (error) {
         loginButtonsSession.errorMessage(error.reason || "Unknown error");
       } else {
@@ -67,9 +66,8 @@ var resetPassword = function () {
 };
 
 Template._resetPasswordDialog.helpers({
-  inResetPasswordFlow: function () {
-    return loginButtonsSession.get('resetPasswordToken');
-  }
+  displayName,
+  inResetPasswordFlow: () => loginButtonsSession.get('resetPasswordToken'),
 });
 
 //
@@ -77,16 +75,13 @@ Template._resetPasswordDialog.helpers({
 //
 
 Template._justResetPasswordDialog.events({
-  'click #just-verified-dismiss-button': function () {
-    loginButtonsSession.set('justResetPassword', false);
-  }
+  'click #just-verified-dismiss-button': () =>
+    loginButtonsSession.set('justResetPassword', false),
 });
 
 Template._justResetPasswordDialog.helpers({
-  visible: function () {
-    return loginButtonsSession.get('justResetPassword');
-  },
-  displayName: displayName
+  visible: () => loginButtonsSession.get('justResetPassword'),
+  displayName,
 });
 
 
@@ -95,30 +90,15 @@ Template._justResetPasswordDialog.helpers({
 // enrollAccountDialog template
 //
 
-Template._enrollAccountDialog.events({
-  'click #login-buttons-enroll-account-button': function () {
-    enrollAccount();
-  },
-  'keypress #enroll-account-password': function (event) {
-    if (event.keyCode === 13)
-      enrollAccount();
-  },
-  'click #login-buttons-cancel-enroll-account': function () {
-    loginButtonsSession.set('enrollAccountToken', null);
-    if (doneCallback)
-      doneCallback();
-  }
-});
-
-var enrollAccount = function () {
+const enrollAccount = () => {
   loginButtonsSession.resetMessages();
-  var password = document.getElementById('enroll-account-password').value;
+  const password = document.getElementById('enroll-account-password').value;
   if (!validatePassword(password))
     return;
 
   Accounts.resetPassword(
     loginButtonsSession.get('enrollAccountToken'), password,
-    function (error) {
+    error => {
       if (error) {
         loginButtonsSession.errorMessage(error.reason || "Unknown error");
       } else {
@@ -129,10 +109,22 @@ var enrollAccount = function () {
     });
 };
 
-Template._enrollAccountDialog.helpers({
-  inEnrollAccountFlow: function () {
-    return loginButtonsSession.get('enrollAccountToken');
+Template._enrollAccountDialog.events({
+  'click #login-buttons-enroll-account-button': enrollAccount,
+  'keypress #enroll-account-password': event => {
+    if (event.keyCode === 13)
+      enrollAccount();
+  },
+  'click #login-buttons-cancel-enroll-account': () => {
+    loginButtonsSession.set('enrollAccountToken', null);
+    if (doneCallback)
+      doneCallback();
   }
+});
+
+Template._enrollAccountDialog.helpers({
+  displayName,
+  inEnrollAccountFlow: () => loginButtonsSession.get('enrollAccountToken'),
 });
 
 
@@ -141,16 +133,13 @@ Template._enrollAccountDialog.helpers({
 //
 
 Template._justVerifiedEmailDialog.events({
-  'click #just-verified-dismiss-button': function () {
-    loginButtonsSession.set('justVerifiedEmail', false);
-  }
+  'click #just-verified-dismiss-button': () =>
+    loginButtonsSession.set('justVerifiedEmail', false),
 });
 
 Template._justVerifiedEmailDialog.helpers({
-  visible: function () {
-    return loginButtonsSession.get('justVerifiedEmail');
-  },
-  displayName: displayName
+  visible: () => loginButtonsSession.get('justVerifiedEmail'),
+  displayName,
 });
 
 
@@ -159,14 +148,13 @@ Template._justVerifiedEmailDialog.helpers({
 //
 
 Template._loginButtonsMessagesDialog.events({
-  'click #messages-dialog-dismiss-button': function () {
-    loginButtonsSession.resetMessages();
-  }
+  'click #messages-dialog-dismiss-button': () =>
+    loginButtonsSession.resetMessages(),
 });
 
 Template._loginButtonsMessagesDialog.helpers({
-  visible: function () {
-    var hasMessage = loginButtonsSession.get('infoMessage') || loginButtonsSession.get('errorMessage');
+  visible: () => {
+    const hasMessage = loginButtonsSession.get('infoMessage') || loginButtonsSession.get('errorMessage');
     return !dropdown() && hasMessage;
   }
 });
@@ -177,42 +165,40 @@ Template._loginButtonsMessagesDialog.helpers({
 //
 
 Template._configureLoginServiceDialog.events({
-  'click .configure-login-service-dismiss-button': function () {
-    loginButtonsSession.set('configureLoginServiceDialogVisible', false);
-  },
-  'click #configure-login-service-dialog-save-configuration': function () {
+  'click .configure-login-service-dismiss-button': () =>
+    loginButtonsSession.set('configureLoginServiceDialogVisible', false),
+  'click #configure-login-service-dialog-save-configuration': () => {
     if (loginButtonsSession.get('configureLoginServiceDialogVisible') &&
         ! loginButtonsSession.get('configureLoginServiceDialogSaveDisabled')) {
       // Prepare the configuration document for this login service
-      var serviceName = loginButtonsSession.get('configureLoginServiceDialogServiceName');
-      var configuration = {
+      const serviceName = loginButtonsSession.get('configureLoginServiceDialogServiceName');
+      const configuration = {
         service: serviceName
       };
 
       // Fetch the value of each input field
-      _.each(configurationFields(), function(field) {
+      configurationFields().forEach(field => {
         configuration[field.property] = document.getElementById(
-          'configure-login-service-dialog-' + field.property).value
+          `configure-login-service-dialog-${field.property}`).value
           .replace(/^\s*|\s*$/g, ""); // trim() doesnt work on IE8;
       });
 
-      Array.prototype.some.call(
-        document.getElementById("configure-login-service-dialog")
-          .getElementsByTagName("input"),
-        function (input) {
-          if (input.getAttribute("name") === "loginStyle" &&
-              input.checked) {
-            configuration.loginStyle = input.value;
-            return true;
-          }
-        }
+      // Replacement of single use of jQuery in this package so we can remove
+      // the dependency
+      const inputs = [].slice.call( // Because HTMLCollections aren't arrays
+        document
+          .getElementById('configure-login-service-dialog')
+          .getElementsByTagName('input')
       );
+
+      configuration.loginStyle =
+        document.querySelector('#configure-login-service-dialog input[name="loginStyle"]:checked').value;
 
       // Configure this login service
       Accounts.connection.call(
-        "configureLoginService", configuration, function (error, result) {
+        "configureLoginService", configuration, (error, result) => {
           if (error)
-            Meteor._debug("Error configuring login service " + serviceName,
+            Meteor._debug(`Error configuring login service ${serviceName}`,
                           error);
           else
             loginButtonsSession.set('configureLoginServiceDialogVisible',
@@ -223,7 +209,7 @@ Template._configureLoginServiceDialog.events({
   // IE8 doesn't support the 'input' event, so we'll run this on the keyup as
   // well. (Keeping the 'input' event means that this also fires when you use
   // the mouse to change the contents of the field, eg 'Cut' menu item.)
-  'input, keyup input': function (event) {
+  'input, keyup input': event => {
     // if the event fired on one of the configuration input fields,
     // check whether we should enable the 'save configuration' button
     if (event.target.id.indexOf('configure-login-service-dialog') === 0)
@@ -234,62 +220,55 @@ Template._configureLoginServiceDialog.events({
 // check whether the 'save configuration' button should be enabled.
 // this is a really strange way to implement this and a Forms
 // Abstraction would make all of this reactive, and simpler.
-var updateSaveDisabled = function () {
-  var anyFieldEmpty = _.any(configurationFields(), function(field) {
-    return document.getElementById(
-      'configure-login-service-dialog-' + field.property).value === '';
-  });
+const updateSaveDisabled = () => {
+  const anyFieldEmpty = configurationFields().reduce((prev, field) =>
+    prev || document.getElementById(
+      `configure-login-service-dialog-${field.property}`
+    ).value === '',
+    false
+  );
 
   loginButtonsSession.set('configureLoginServiceDialogSaveDisabled', anyFieldEmpty);
 };
 
 // Returns the appropriate template for this login service.  This
 // template should be defined in the service's package
-Template._configureLoginServiceDialog.templateForService = function(serviceName) {
+Template._configureLoginServiceDialog.templateForService = serviceName => {
   serviceName = serviceName || loginButtonsSession.get('configureLoginServiceDialogServiceName');
   // XXX Service providers should be able to specify their configuration
   // template name.
-  return Template['configureLoginServiceDialogFor' +
-                  (serviceName === 'meteor-developer' ?
+  return Template[`configureLoginServiceDialogFor${
+                  serviceName === 'meteor-developer' ?
                    'MeteorDeveloper' :
-                   capitalize(serviceName))];
+                   capitalize(serviceName)}`];
 };
 
-var configurationFields = function () {
-  var template = Template._configureLoginServiceDialog.templateForService();
+const configurationFields = () => {
+  const template = Template._configureLoginServiceDialog.templateForService();
   return template.fields();
 };
 
 Template._configureLoginServiceDialog.helpers({
-  configurationFields: function () {
-    return configurationFields();
-  },
-  visible: function () {
-    return loginButtonsSession.get('configureLoginServiceDialogVisible');
-  },
-  configurationSteps: function () {
-    // renders the appropriate template
-    return Template._configureLoginServiceDialog.templateForService();
-  },
-  saveDisabled: function () {
-    return loginButtonsSession.get('configureLoginServiceDialogSaveDisabled');
-  }
+  configurationFields,
+  visible: () => loginButtonsSession.get('configureLoginServiceDialogVisible'),
+  // renders the appropriate template
+  configurationSteps: () =>
+    Template._configureLoginServiceDialog.templateForService(),
+  saveDisabled: () =>
+    loginButtonsSession.get('configureLoginServiceDialogSaveDisabled'),
 });
 
 // XXX from http://epeli.github.com/underscore.string/lib/underscore.string.js
-var capitalize = function(str){
+const capitalize = str => {
   str = str == null ? '' : String(str);
   return str.charAt(0).toUpperCase() + str.slice(1);
 };
 
 Template._configureLoginOnDesktopDialog.helpers({
-  visible: function () {
-    return loginButtonsSession.get('configureOnDesktopVisible');
-  }
+  visible: () => loginButtonsSession.get('configureOnDesktopVisible'),
 });
 
 Template._configureLoginOnDesktopDialog.events({
-  'click #configure-on-desktop-dismiss-button': function () {
-    loginButtonsSession.set('configureOnDesktopVisible', false);
-  }
+  'click #configure-on-desktop-dismiss-button': () =>
+    loginButtonsSession.set('configureOnDesktopVisible', false),
 });