+++ packages/accounts-ui/package.js
@@ -1,9 +1,9 @@
 Package.describe({
   summary: "Simple templates to add login widgets to an app",
-  version: "1.3.0"
+  version: "1.3.1",
 });
 
-Package.onUse(function (api) {
+Package.onUse(api => {
   // Export Accounts (etc) to packages using this one.
   api.imply('accounts-base', ['client', 'server']);
   api.use('accounts-ui-unstyled', 'client');