+++ packages/binary-heap/binary-heap.js
@@ -1,0 +1,3 @@
+export { MaxHeap } from './max-heap.js';
+export { MinHeap } from './min-heap.js';
+export { MinMaxHeap } from './min-max-heap.js';