+++ packages/meetup-oauth/meetup_client.js
@@ -4,14 +4,14 @@ Meetup = {};
 // @param credentialRequestCompleteCallback {Function} Callback function to call on
 //   completion. Takes one argument, credentialToken on success, or Error on
 //   error.
-Meetup.requestCredential = function (options, credentialRequestCompleteCallback) {
+Meetup.requestCredential = (options, credentialRequestCompleteCallback) => {
   // support both (options, callback) and (callback).
   if (!credentialRequestCompleteCallback && typeof options === 'function') {
     credentialRequestCompleteCallback = options;
     options = {};
   }
 
-  var config = ServiceConfiguration.configurations.findOne({service: 'meetup'});
+  const config = ServiceConfiguration.configurations.findOne({service: 'meetup'});
   if (!config) {
     credentialRequestCompleteCallback && credentialRequestCompleteCallback(
       new ServiceConfiguration.ConfigError());
@@ -22,32 +22,32 @@ Meetup.requestCredential = function (options, credentialRequestCompleteCallback)
   // parameter when redirecting back to the client, so we use
   // `Random.id()` here (alphanumerics) instead of `Random.secret()`
   // (base 64 characters).
-  var credentialToken = Random.id();
+  const credentialToken = Random.id();
 
-  var scope = (options && options.requestPermissions) || [];
-  var flatScope = _.map(scope, encodeURIComponent).join('+');
+  const scope = (options && options.requestPermissions) || [];
+  const flatScope = scope.map(encodeURIComponent).join('+');
 
-  var loginStyle = OAuth._loginStyle('meetup', config, options);
+  const loginStyle = OAuth._loginStyle('meetup', config, options);
 
-  var loginUrl =
+  const loginUrl =
         'https://secure.meetup.com/oauth2/authorize' +
-        '?client_id=' + config.clientId +
+        `?client_id=${config.clientId}` +
         '&response_type=code' +
-        '&scope=' + flatScope +
-        '&redirect_uri=' + OAuth._redirectUri('meetup', config) +
-        '&state=' + OAuth._stateParam(loginStyle, credentialToken, options && options.redirectUrl);
+        `&scope=${flatScope}` +
+        `&redirect_uri=${OAuth._redirectUri('meetup', config)}` +
+        `&state=${OAuth._stateParam(loginStyle, credentialToken, options && options.redirectUrl)}`;
 
   // meetup box gets taller when permissions requested.
-  var height = 620;
-  if (_.without(scope, 'basic').length)
+  let height = 620;
+  if (Object.prototype.hasOwnProperty.call(scope, 'basic') ? scope.length - 1 : scope.length)
     height += 130;
 
   OAuth.launchLogin({
     loginService: "meetup",
-    loginStyle: loginStyle,
-    loginUrl: loginUrl,
-    credentialRequestCompleteCallback: credentialRequestCompleteCallback,
-    credentialToken: credentialToken,
-    popupOptions: {width: 900, height: height}
+    loginStyle,
+    loginUrl,
+    credentialRequestCompleteCallback,
+    credentialToken,
+    popupOptions: { width: 900, height },
   });
 };