+++ packages/meetup-oauth/package.js
@@ -1,13 +1,12 @@
 Package.describe({
   summary: 'Meetup OAuth flow',
-  version: '1.0.2'
+  version: '1.0.3'
 });
 
-Package.onUse(function (api) {
+Package.onUse(api => {
   api.use('oauth2', ['client', 'server']);
   api.use('oauth', ['client', 'server']);
-  api.use('http', ['server']);
-  api.use('underscore', ['client', 'server']);
+  api.use('http', 'server');
   api.use('random', 'client');
   api.use('service-configuration', ['client', 'server']);