+++ packages/oauth/oauth_browser.js
@@ -8,20 +8,21 @@
 //   arguments.
 // @param dimensions {optional Object(width, height)} The dimensions of
 //   the popup. If not passed defaults to something sane.
-OAuth.showPopup = function (url, callback, dimensions) {
+OAuth.showPopup = (url, callback, dimensions) => {
   // default dimensions that worked well for facebook and google
-  var popup = openCenteredPopup(
+  const popup = openCenteredPopup(
     url,
     (dimensions && dimensions.width) || 650,
     (dimensions && dimensions.height) || 331
   );
 
-  var checkPopupOpen = setInterval(function() {
+  const checkPopupOpen = setInterval(() => {
+    let popupClosed;
     try {
       // Fix for #328 - added a second test criteria (popup.closed === undefined)
       // to humour this Android quirk:
       // http://code.google.com/p/android/issues/detail?id=21061
-      var popupClosed = popup.closed || popup.closed === undefined;
+      popupClosed = popup.closed || popup.closed === undefined;
     } catch (e) {
       // For some unknown reason, IE9 (and others?) sometimes (when
       // the popup closes too quickly?) throws "SCRIPT16386: No such
@@ -37,29 +38,29 @@ OAuth.showPopup = function (url, callback, dimensions) {
   }, 100);
 };
 
-var openCenteredPopup = function(url, width, height) {
-  var screenX = typeof window.screenX !== 'undefined'
+const openCenteredPopup = function(url, width, height) {
+  const screenX = typeof window.screenX !== 'undefined'
         ? window.screenX : window.screenLeft;
-  var screenY = typeof window.screenY !== 'undefined'
+  const screenY = typeof window.screenY !== 'undefined'
         ? window.screenY : window.screenTop;
-  var outerWidth = typeof window.outerWidth !== 'undefined'
+  const outerWidth = typeof window.outerWidth !== 'undefined'
         ? window.outerWidth : document.body.clientWidth;
-  var outerHeight = typeof window.outerHeight !== 'undefined'
+  const outerHeight = typeof window.outerHeight !== 'undefined'
         ? window.outerHeight : (document.body.clientHeight - 22);
   // XXX what is the 22?
 
   // Use `outerWidth - width` and `outerHeight - height` for help in
   // positioning the popup centered relative to the current window
-  var left = screenX + (outerWidth - width) / 2;
-  var top = screenY + (outerHeight - height) / 2;
-  var features = ('width=' + width + ',height=' + height +
-                  ',left=' + left + ',top=' + top + ',scrollbars=yes');
+  const left = screenX + (outerWidth - width) / 2;
+  const top = screenY + (outerHeight - height) / 2;
+  const features = (`width=${width},height=${height}` +
+                  `,left=${left},top=${top},scrollbars=yes'`);
 
-  var newwindow = window.open(url, 'Login', features);
+  const newwindow = window.open(url, 'Login', features);
 
   if (typeof newwindow === 'undefined') {
     // blocked by a popup blocker maybe?
-    var err = new Error("The login popup was blocked by the browser");
+    const err = new Error("The login popup was blocked by the browser");
     err.attemptedUrl = url;
     throw err;
   }