+++ packages/oauth/oauth_common.js
@@ -1,6 +1,8 @@
+import url from 'url';
+
 OAuth._storageTokenPrefix = "Meteor.oauth.credentialSecret-";
 
-OAuth._redirectUri = function (serviceName, config, params, absoluteUrlOptions) {
+OAuth._redirectUri = (serviceName, config, params, absoluteUrlOptions) => {
   // XXX COMPAT WITH 0.9.0
   // The redirect URI used to have a "?close" query argument.  We
   // detect whether we need to be backwards compatible by checking for
@@ -8,26 +10,26 @@ OAuth._redirectUri = function (serviceName, config, params, absoluteUrlOptions) 
   // code which had the "?close" argument.
   // This logic is duplicated in the tool so that the tool can do OAuth
   // flow with <= 0.9.0 servers (tools/auth.js).
-  var query = config.loginStyle ? null : "close";
+  const query = config.loginStyle ? null : "close";
 
   // Clone because we're going to mutate 'params'. The 'cordova' and
   // 'android' parameters are only used for picking the host of the
   // redirect URL, and not actually included in the redirect URL itself.
-  var isCordova = false;
-  var isAndroid = false;
+  let isCordova = false;
+  let isAndroid = false;
   if (params) {
-    params = _.clone(params);
+    params = { ...params };
     isCordova = params.cordova;
     isAndroid = params.android;
     delete params.cordova;
     delete params.android;
-    if (_.isEmpty(params)) {
+    if (Object.keys(params).length === 0) {
       params = undefined;
     }
   }
 
   if (Meteor.isServer && isCordova) {
-    var rootUrl = process.env.MOBILE_ROOT_URL ||
+    let rootUrl = process.env.MOBILE_ROOT_URL ||
           __meteor_runtime_config__.ROOT_URL;
 
     if (isAndroid) {
@@ -36,8 +38,7 @@ OAuth._redirectUri = function (serviceName, config, params, absoluteUrlOptions) 
       // XXX Maybe we should put this in a separate package or something
       // that is used here and by boilerplate-generator? Or maybe
       // `Meteor.absoluteUrl` should know how to do this?
-      var url = Npm.require("url");
-      var parsedRootUrl = url.parse(rootUrl);
+      const parsedRootUrl = url.parse(rootUrl);
       if (parsedRootUrl.hostname === "localhost") {
         parsedRootUrl.hostname = "10.0.2.2";
         delete parsedRootUrl.host;
@@ -45,15 +46,16 @@ OAuth._redirectUri = function (serviceName, config, params, absoluteUrlOptions) 
       rootUrl = url.format(parsedRootUrl);
     }
 
-    absoluteUrlOptions = _.extend({}, absoluteUrlOptions, {
+    absoluteUrlOptions = { 
+      ...absoluteUrlOptions,
       // For Cordova clients, redirect to the special Cordova root url
       // (likely a local IP in development mode).
-      rootUrl: rootUrl
-    });
+      rootUrl,
+    };
   }
 
   return URL._constructUrl(
-    Meteor.absoluteUrl('_oauth/' + serviceName, absoluteUrlOptions),
+    Meteor.absoluteUrl(`_oauth/${serviceName}`, absoluteUrlOptions),
     query,
     params);
 };